// forum-ui/src/app/api/models/post-model.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface PostModel {
  authorId?: string;
  category?: string | null;
  content?: string | null;
  createdAt?: string;
  id?: string;
  isPublished?: boolean;
  tagNames?: Array<string> | null;
  title?: string | null;
  updatedAt?: string;
}